Move RunData type to prevent cyclic dependency
[MAILPOET-5436]
